Adding updated tests
