--- controls/controls.py
@@ -1,2 +1,0 @@
-class AmityControls():
-    pass
