+++ files/person.txt
@@ -1,0 +1,7 @@
+OLUWAFEMI SULE FELLOW Y
+DOMINIC WALTERS STAFF N
+SIMON PATTERSON FELLOW Y
+MARI LAWRENCE FELLOW Y
+LEIGH RILEY STAFF N
+TANA LOPEZ FELLOW Y
+KELLY McGUIRE STAFF N
